Simplify tsdown package version resolution
