+++ configs/tsdown/plugins.test.ts
@@ -1,4 +1,3 @@
-import fs from 'node:fs';
 import { fileURLToPath } from 'node:url';
 import { describe, expect, it } from 'vitest';
 import { createBasePlugin } from './plugins.ts';
@@ -10,24 +9,19 @@ describe('tsdown base plugin', () => {
       throw new Error('Expected the tsdown base plugin to define a transform hook.');
     }
 
-    const moduleId = fileURLToPath(new URL(
-      '../../packages/js/src/lib/hexclave-app/apps/implementations/common.ts',
-      import.meta.url,
-    ));
-    const packageJson: { name: string; version: string } = JSON.parse(fs.readFileSync(
-      fileURLToPath(new URL('../../packages/js/package.json', import.meta.url)),
-      'utf-8',
-    ));
+    if (typeof plugin.buildStart !== 'function') {
+      throw new Error('Expected the tsdown base plugin to define a buildStart hook.');
+    }
+    const packageRoot = fileURLToPath(new URL('../../packages/js', import.meta.url));
+    Reflect.apply(plugin.buildStart, undefined, [{ cwd: packageRoot }]);
     const result = await Reflect.apply(plugin.transform, undefined, [
       'const clientVersion = "STACK_COMPILE_TIME_CLIENT_PACKAGE_VERSION_SENTINEL";',
-      moduleId,
+      `${packageRoot}/src/lib/hexclave-app/apps/implementations/common.ts`,
       {},
     ]);
 
     expect(result).not.toBeNull();
-    expect(typeof result === 'string' ? result : result?.code).toContain(
-      `js ${packageJson.name}@${packageJson.version}`,
-    );
+    expect(typeof result === 'string' ? result : result?.code).toMatch(/js @hexclave\/js@\d+\.\d+\.\d+/);
     expect(typeof result === 'string' ? result : result?.code).not.toContain('@hexclave/monorepo@0.0.0');
   });
 });
